Abandoned behaviour class
Behaviour was meant to have various states to keep track of
different states of being. It was supposed to switch from one
to another. After the LLM was introduced, this just isn't
necessary, so ripped it out.

Eventually behaviour will be folded into broca. It's just a
meaningless middleman right now.